Added a ChatBot class
We were testing the idea of using a chat bot that would save all the chats from the room and send them to any new person who enters the room.
We quickly realized that this chat bot however, won't be able to stay online as long as we need it to.
that's why we didn't finish the chat bot class

Co-Authored-By: Mohammad Bashitul Islam <82585669+Casrial77@users.noreply.github.com>